# backend/api-first/openapi.yaml
openapi: 3.1.0
info:
  title: KOSMO API
  version: "v1"
  description: >-
    API de KOSMO - plataforma low-code de spec-driven development (SDD) asistida por IA.
    Ciclo completo: Descubrimiento -> Características -> Requisitos -> Modelo -> Código.

    Convenciones:
    - Autenticación PKCE (RFC 7636) con JWT RS256 (bearerAuth).
    - Errores RFC 7807 (application/problem+json) con type URN, trace_id y violations, o OAuthErrorResponse (RFC 6749) / HttpErrorResponse.
    - Fechas ISO-8601 UTC con sufijo Z.
    - Las fases siguen el orden Descubrimiento -> Características -> Requisitos -> Modelo -> Código.
      La consistencia se evalúa exclusivamente hacia la derecha (downstream).
  contact:
    name: Equipo KOSMO
    email: dev@kosmo.app
    url: https://github.com/CesarPantoja1/KOSMO
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local — desarrollo en máquina del programador
  - url: https://api-dev.kosmo.app
    description: Desarrollo — entorno de integración continua
  - url: https://api.kosmo.app
    description: Producción — tráfico real de usuarios
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/projects/{project_id}/features:
    post:
      tags: [features]
      summary: Generar caracteristicas del producto con IA
      description: Genera caracteristicas de alto nivel evaluando el documento de descubrimiento del proyecto.
      operationId: features_generate
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Caracteristicas generadas exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [features]
      summary: Listar caracteristicas del proyecto
      description: Devuelve todas las caracteristicas asociadas a un proyecto.
      operationId: features_list
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Lista de caracteristicas del proyecto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/projects/{project_id}/features/suggest:
    post:
      tags: [features]
      summary: Sugerir nuevas caracteristicas con IA
      description: Sugiere caracteristicas adicionales basadas en el documento de descubrimiento evitando duplicados.
      operationId: features_suggest
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Sugerencias de caracteristicas generadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureSuggestionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/projects/{project_id}/features/manual:
    post:
      tags: [features]
      summary: Crear caracteristica manualmente
      description: >-
        Crea una nueva caracteristica. Si no se indica origin, la IA lo deriva del descubrimiento
        y verifica coherencia. Ante inconsistencia devuelve is_saved=false con la razon.
      operationId: features_create_manual
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacteristicRequest"
      responses:
        "200":
          description: Caracteristica creada exitosamente o rechazada por inconsistencia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCharacteristicResult"
        "400":
          description: Datos invalidos (titulo vacio o excesivo, descripcion muy larga)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/projects/{project_id}/features/{feature_id}/manual:
    put:
      tags: [features]
      summary: Editar caracteristica manualmente
      description: >-
        Edita una caracteristica existente. Si los cambios contradicen flagrantemente el descubrimiento,
        el guardado es rechazado devolviendo 409 Conflict.
      operationId: features_edit_manual
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditFeatureManualRequest"
      responses:
        "200":
          description: Caracteristica editada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Contradiccion detectada con el documento fuente
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/projects/{project_id}/features/save:
    post:
      tags: [features]
      summary: Guardar caracteristicas seleccionadas
      description: Guarda las caracteristicas seleccionadas por el usuario desde las sugerencias de la IA.
      operationId: features_save_selected
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSelectedFeaturesRequest"
      responses:
        "201":
          description: Caracteristicas guardadas exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/projects/{project_id}/features/{feature_id}:
    delete:
      tags: [features]
      summary: Eliminar caracteristica
      description: Elimina una caracteristica y todos sus artefactos asociados (requisitos, diagramas).
      operationId: features_delete
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FeatureId"
      responses:
        "200":
          description: Caracteristica eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
                  feature_id:
                    type: string
                required: [status, feature_id]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/projects/{project_id}/features/{feature_id}/consistency/check:
    post:
      tags: [features]
      summary: Verificar consistencia antes de guardado manual
      description: >-
        Verifica que el contenido editado no contradiga el descubrimiento invocando al agente IA.
      operationId: features_check_consistency
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FeatureId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckConsistencyRequestView"
      responses:
        "200":
          description: Resultado de la verificacion de consistencia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InconsistencyResultView"
        "404":
          $ref: "#/components/responses/NotFound"

  # ══════════════════════════════ FEATURE CHAT ══════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT RS256 emitido por /api/v1/auth/token o /api/v1/auth/refresh
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
      description: Identificador del proyecto (prj_ + ULID)
    FeatureId:
      name: feature_id
      in: path
      required: true
      schema:
        type: string
      description: Identificador o slug de la característica
...
  responses:
    Unauthorized:
      description: Token de acceso ausente, invalido, expirado o revocado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HttpErrorResponse"
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    BadGateway:
      description: Error al invocar el servicio de IA o LLM
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HttpErrorResponse"
...
  schemas:
    ProblemDetail:
      type: object
      description: Error RFC 7807 (application/problem+json)
      properties:
        type:
          type: string
          description: URN del error (ej. urn:kosmo:projects:not-found)
          example: urn:kosmo:projects:not-found
        title:
          type: string
          description: Titulo corto del error
        status:
          type: integer
          description: Codigo de estado HTTP
        detail:
          type: string
          description: Descripcion detallada del error
        instance:
          type: string
          description: Ruta o recurso que origino el error
        trace_id:
          type: string
          description: Identificador de correlacion distribuida (X-Trace-Id)
        violations:
          type: array
          items:
            type: object
            properties:
              loc:
                type: string
                description: Ruta al campo invalido
              msg:
                type: string
                description: Mensaje de error de validacion
              input:
                description: Valor recibido que causo la violacion
      required: [type, title, status, detail]
    HttpErrorResponse:
      type: object
      description: Respuesta de error generica para errores HTTP (4xx/5xx)
      properties:
        detail:
          type: string
          description: Mensaje descriptivo del error
          example: Recurso no encontrado.
      required: [detail]

    # ── Auth ──
...
    FeatureResponse:
      type: object
      properties:
        id:
          type: string
          description: Identificador unico de la caracteristica
          example: feat_01HXYAZABCDEFGHIJKLMNOP
        project_id:
          type: string
          example: prj_01HXYAZABCDEFGHIJKLMNOP
        number:
          type: integer
          example: 1
        title:
          type: string
          example: Gestion de catalogo de productos
        slug:
          type: string
          example: gestion-de-catalogo-de-productos
        description:
          type: string
          example: Permite a los usuarios administrar el catalogo de productos.
        origin:
          type: string
          example: Se deriva de la seccion Capacidades del descubrimiento.
        display_id:
          type: string
          example: C01
        warnings:
          type: array
          items:
            type: string
          description: Advertencias sobre solapamientos funcionales o inconsistencias menores
          default: []
      required: [id, project_id, number, title, slug, description, origin, display_id]
    FeatureSuggestionItem:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        origin:
          type: string
          default: ""
      required: [title, description]
    FeatureSuggestionResponse:
      type: object
      properties:
        number:
          type: integer
          example: 1
        title:
          type: string
          example: Modulo de facturacion electronica
        description:
          type: string
          example: Generacion y envio de comprobantes tributarios.
        origin:
          type: string
          example: Derivado de Reglas de Negocio en Descubrimiento.
      required: [number, title, description, origin]
    SaveSelectedFeaturesRequest:
      type: object
      properties:
        features:
          type: array
          items:
            $ref: "#/components/schemas/FeatureSuggestionItem"
      required: [features]
    CreateCharacteristicRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 50
          description: Titulo de la caracteristica (maximo 50 caracteres)
          example: Gestion de catalogo de productos
        description:
          type: string
          maxLength: 500
          default: ""
          description: Descripcion de la caracteristica (maximo 500 caracteres)
          example: Permite a los usuarios administrar el catalogo de productos.
        origin:
          type: string
          default: ""
          description: Origen y trazabilidad. Si se omite, la IA lo deriva.
      required: [title]
    EditFeatureManualRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 50
          description: Nuevo titulo de la caracteristica
        description:
          type: string
          minLength: 0
          maxLength: 500
          description: Nueva descripcion de la caracteristica
      required: [title, description]
    CreateCharacteristicResult:
      type: object
      properties:
        is_saved:
          type: boolean
          description: True si la caracteristica fue guardada
        feature:
          $ref: "#/components/schemas/FeatureResponse"
          description: Presente solo si is_saved=true
        origin:
          type: string
        is_consistent:
          type: boolean
          description: False si se detecto contradiccion con el descubrimiento
        inconsistency_reason:
          type: string
          nullable: true
          description: Presente solo si is_saved=false por inconsistencia
        warnings:
          type: array
          items:
            type: string
          description: Advertencias sobre solapamientos funcionales o inconsistencias menores
          default: []
      required: [is_saved, origin, is_consistent]
    CheckConsistencyRequestView:
      type: object
      additionalProperties: false
      properties:
        content:
          type: object
          description: Contenido nuevo a validar contra el documento fuente
      required: [content]
    InconsistencyResultView:
      type: object
      properties:
        is_consistent:
          type: boolean
          description: False si se detecto contradiccion
        reason:
          type: string
          nullable: true
          description: Motivo de la inconsistencia
        conflicting_section:
          type: string
          nullable: true
          description: Seccion del documento fuente con conflicto
      required: [is_consistent]

    # ── Chat ──
...